re-use account fields component in account header
