+++ app/javascript/mastodon/components/hover_card_account.tsx
@@ -23,6 +23,8 @@ import { domain } from 'mastodon/initial_state';
 import { getAccountHidden } from 'mastodon/selectors/accounts';
 import { useAppSelector, useAppDispatch } from 'mastodon/store';
 
+import { useLinks } from '../hooks/useLinks';
+
 export const HoverCardAccount = forwardRef<
   HTMLDivElement,
   { accountId?: string }
@@ -63,6 +65,8 @@ export const HoverCardAccount = forwardRef<
     hasRelationshipLoaded &&
     !isMutual &&
     !isFollower;
+
+  const handleClick = useLinks();
 
   return (
     <div
@@ -105,11 +109,14 @@ export const HoverCardAccount = forwardRef<
                   accountId={account.id}
                   className='hover-card__bio'
                 />
-                <AccountFields
-                  fields={account.fields}
-                  limit={2}
-                  extraEmojis={account.emojis}
-                />
+
+                <div className='account-fields' onClickCapture={handleClick}>
+                  <AccountFields
+                    fields={account.fields.take(2)}
+                    emojis={account.emojis}
+                  />
+                </div>
+
                 {note && note.length > 0 && (
                   <dl className='hover-card__note'>
                     <dt className='hover-card__note-label'>
